Embedding SciChart to MultiLineLabels component
